requirements.txt

django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
hiredis
python-dotenv
drf-yasg